split test_plugins.py
